+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Allies warn you when an enemy fleet masses in orbit above one of your planets
 - Allies warn you when the enemy reaches a planet your team holds
 - Allies call for help when their army on a planet is being destroyed
+- Support the Exiles faction
 - Fix issues with Legion Expansion unit detection
 - Fix vehicles (not tanks) and scouts not counting towards an ally's army size
 - Fix factories counting towards an ally's army size